use uint16 instead of ints
~~~
goos: linux
goarch: arm64
pkg: codeberg.org/miekg/dns/internal/pack
       │   old.txt   │               new.txt               │
       │   sec/op    │   sec/op     vs base                │
Name-8   35.51n ± 2%   30.18n ± 2%  -15.02% (p=0.000 n=10)
~~~

We don't need full blown ints here, we know the maxium sizes.

+++ dnsutil/shared.go
@@ -143,7 +143,7 @@ func IsName(s string) bool {
 	// XXX: The logic in this function was copied from pack.Name and should be kept in sync with that function.
 
 	const lenmsg = 256
-	ls := len(s)
+	ls := uint16(len(s))
 
 	if ls == 1 && s[0] == '.' {
 		return true
@@ -154,11 +154,11 @@ func IsName(s string) bool {
 	}
 
 	var (
-		off    int
-		begin  int
+		off    uint16
+		begin  uint16
 		escape bool
 	)
-	for i := 0; i < ls; i++ {
+	for i := uint16(0); i < ls; i++ {
 		switch s[i] {
 		case '\\':
 			escape = !escape
